updater/requirements.txt
loguru==0.6.0
psycopg2-binary==2.9.4